Move parameter examples to schema
Having x-examples on an `in: body` parameter caused swagger-codegen to
fail (swagger-api/swagger-codegen issue 4442).

Although having the example on the schema is ambiguous for types which
are used in multiple contexts (e.g. PUT and POST or multiple endpoints)
it is preferred by Dredd so we'll use it for now.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>