Define StringValueObject
A type with a single string-valued "value" property is common in the
API.  Give this type a name.

Note that some uses had minLength: 1, but this had not been tested in
the REST API.  So remove it for now and re-add if it turns out to be a
real constraint.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>