Expand non-polymorphic NotificationSettings type
In the non-polymorphic version, the NotificationSettings type must be
able to represent the settings for any of the notification types.
Expand the definition to do this and add some documentation about which
properties correspond to which notification types.

I am very glad none of the types used the same property name with
different types!

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>