Factor out Timestamped schema
The Timestamped schema is a common interface shared by objects which
maintain created and updated timestamps.

Ideally code generators would treat this as an interface or base class
which could be used in code, although they currently do not.  For now it
is used to document a shared convention and reduce redundant
definitions.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>