Change versioning scheme
Keep the major version at 0, both to signal semver API instability and
so that any numbered version introduced later can compare as larger.

Use the minor number as the API change date.  This describes the version
of the AppVeyor REST API being described.

Use the patch number for iterations of the OpenAPI/Swagger description
which do not reflect a change in the REST API.  Such changes may reflect
a change in any generated programmatic API, so should not be depended
upon without testing.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>