Move query parameters in to GET operation object
Since these query parameters are unlikely to be used on any
POST/PUT/DELETE method that is added for the same path, move them into
the GET operation.

This shouldn't change the generated API, but if it does it's better to
do it now before people start using this definition.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>