Add some property descriptions
Document some property behavior which may not be obvious.

Note that the properties which are missing from the addProject response
could be made more explicit (and compiler-enforced) by refactoring into
a separate base schema.  However, given the pain this causes in some
strictly typed languages (C#, Java) due to lack of inheritance in the
generated classes the tradeoff doesn't seem worth it.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>