Update description for recordsNumber
Today the documentation was updated to indicate 20 as the maximum value
for recordsNumber.  In my testing, higher values were accepted (although
I don't have a project with that many deployments to test whether they
are returned).  It also indicated that values <= 5 caused 500 Internal
Server Error.  Document both, and that other values have failed (see
appveyor-swagger-java for examples - 10, 11, and 12 have worked or
failed at different times) and the heuristic that I currently use.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>