Add inviteUser, getUserInvitations, cancelUserInvitation
As discussed in appveyor/ci#2892, /users/invitations is the preferred
API for adding users/collaborators to a team.  Add support for inviting
a user, getting the pending invitations, and canceling an invitation.

Note: UserInvitationModel is the name from the getUserInvitations
response XML.  InviteUserRequest fits the naming scheme of other request
objects exposed via XML.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>