openapi.yaml: Convert spec from Swagger to OAS3
Convert swagger-polymorphic.yaml to version 3 of the OpenAPI
Specification and remove the (now unnecessary) swagger.yaml.

The plan is to automatically generate the user-level (v2),
non-polymorphic, and Swagger/OAS2 versions of the specification from
this OAS3 spec.

Note: validate-oas is removed from validate script since oas does not
support OpenAPI 3.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>