openapi.yaml: Add discriminator mappings
With OAS3, schema names are no longer required to match the
discriminator value (which prevented use on both DeploymentProvider and
NotificationProviderSettings due to a conflict over the Webhook schema
name and prevented use on NotificationSettings due to names which are
not valid for schemas).

Use reasonable schema names (from the XML where known) and add mappings
values to schema names.

Note: Make $type required in NotificationSettings, since discriminator
properties must be required.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>