Specify v2 (user-level) API in YAML file
Since it is only possible to get v2 tokens (as far as I can see), this
will be the currently supported version of the API going forward.  Old
versions will be generated as necessary for backwards compatibility.

Note: Rename `apiToken` to `apiTokenV2` and document `"v2."` prefix
since this a v2 token is required for the operations specified.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>